requirements.txt
absl-py==1.4.0
accelerate==0.9.0
aiohttp==3.8.4
aiosignal==1.2.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
datasets==2.4.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.4.1
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.3
google-auth==2.17.1
google-auth-oauthlib==0.4.6
grpcio==1.48.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
huggingface-hub==0.4.0
idna==3.4
idna-ssl==1.1.0
imageio==2.15.0
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
json5==0.9.11
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.1
littleutils==0.2.2
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
multidict==5.2.0
multiprocess==0.70.12.2
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
networkx==2.5.1
notebook==6.4.10
numpy==1.19.5
nvidia-smi==0.1.3
oauthlib==3.2.2
opencv-python==4.5.3.56
openpyxl==3.0.10
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pyDeprecate==0.3.1
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.18.0
pytest==7.0.1
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2022.7.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==25.0.0
regex==2023.8.8
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.17.0
rfc3986==1.5.0
rsa==4.9
sacremoses==0.0.53
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sorcery==0.2.2
stream-inflate==0.0.14
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tokenizers==0.12.1
tomli==1.2.3
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchmetrics==0.8.2
torchvision==0.11.3+cu113
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
transformers==4.18.0
typing_extensions==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
wrapt==1.15.0
xlrd==2.0.1
xxhash==3.2.0
yarl==1.7.2
zipp==3.6.0
